requirements.txt
# Core ML Libraries
ultralytics==8.0.227        # YOLOv8
torch>=2.0.0                # PyTorch
torchvision>=0.15.0         # Vision utilities
opencv-python-headless==4.8.1.78  # OpenCV (no GUI)
pillow==10.1.0              # Image processing
numpy==1.26.2               # Numerical computing

# API Server
fastapi==0.104.1            # REST API framework
uvicorn[standard]==0.24.0   # ASGI server
python-multipart==0.0.6     # File upload support

# Data Processing
pandas==2.1.3               # Data manipulation
scikit-learn==1.3.2         # ML utilities
tqdm==4.66.1                # Progress bars

# Jupyter (optional)
jupyter==1.0.0              # Notebook
matplotlib==3.8.2           # Plotting
seaborn==0.13.0             # Statistical plots

# Development
pytest==7.4.3               # Testing
black==23.11.0              # Code formatting
flake8==6.1.0               # Linting
